+++ packages/vscode-extension/src/chat-view-provider.ts
@@ -1,9 +1,8 @@
 import * as vscode from 'vscode';
-import axios, { AxiosHeaders } from 'axios';
+import axios from 'axios';
 import { RawAxiosRequestHeaders } from 'axios/index';
 
 export default class ChatViewProvider implements vscode.WebviewViewProvider {
-
   private webView?: vscode.WebviewView;
   private requestAbortController;
 
@@ -14,8 +13,6 @@ export default class ChatViewProvider implements vscode.WebviewViewProvider {
 
   public resolveWebviewView(
     webviewView: vscode.WebviewView,
-    context: vscode.WebviewViewResolveContext,
-    _token: vscode.CancellationToken,
   ) {
     this.webView = webviewView;
 
@@ -23,12 +20,21 @@ export default class ChatViewProvider implements vscode.WebviewViewProvider {
       // Allow scripts in the webview
       enableScripts: true,
 
-      localResourceRoots: [
-        this.context.extensionUri,
-      ],
+      localResourceRoots: [this.context.extensionUri],
     };
 
     webviewView.webview.html = this.getWebviewHtml(webviewView.webview);
+
+    const {
+      apiBaseUrl,
+      apiKey,
+    } = this.getCredentials();
+
+    if (!apiBaseUrl || !apiKey) {
+      this.webView?.webview.postMessage({
+        type: 'addSetupMessage',
+      });
+    }
 
     webviewView.webview.onDidReceiveMessage(data => {
       switch (data.type) {
@@ -44,13 +50,26 @@ export default class ChatViewProvider implements vscode.WebviewViewProvider {
           this.requestAbortController?.abort();
           break;
         }
+        case 'goToExtensionSettings': {
+          vscode.commands.executeCommand('workbench.action.openSettings', `@ext:${this.context.extension.id}`);
+          break;
+        }
       }
     });
   }
 
+  private getCredentials() {
+    return {
+      apiBaseUrl: vscode.workspace.getConfiguration('poly').get('apiBaseUrl'),
+      apiKey: vscode.workspace.getConfiguration('poly').get('apiKey'),
+    };
+  }
+
   private async sendPolyQuestionRequest(message: string) {
-    const apiBaseUrl = vscode.workspace.getConfiguration('poly').get('apiBaseUrl');
-    const apiKey = vscode.workspace.getConfiguration('poly').get('apiKey');
+    const {
+      apiBaseUrl,
+      apiKey,
+    } = this.getCredentials();
 
     if (!apiBaseUrl || !apiKey) {
       vscode.window.showErrorMessage('Please set the API base URL and API key in the extension settings.');
@@ -86,10 +105,12 @@ export default class ChatViewProvider implements vscode.WebviewViewProvider {
       console.error(error);
       this.webView?.webview.postMessage({
         type: 'addResponseTexts',
-        value: [{
-          type: 'error',
-          value: error.response?.data?.message || error.message,
-        }],
+        value: [
+          {
+            type: 'error',
+            value: error.response?.data?.message || error.message,
+          },
+        ],
       });
     }
   }
@@ -120,7 +141,7 @@ export default class ChatViewProvider implements vscode.WebviewViewProvider {
         command,
       }, {
         headers: {
-          'Authorization': `Bearer ${apiKey}`,
+          Authorization: `Bearer ${apiKey}`,
         } as RawAxiosRequestHeaders,
       });
     } catch (error) {
@@ -177,7 +198,7 @@ export default class ChatViewProvider implements vscode.WebviewViewProvider {
             </button>
           </div>
         </div>
-		    <script src='${mainJs}'></script>
+        <script src='${mainJs}'></script>
       </body>
       </html>`
     );