+++ test/mocks/event.service.ts
@@ -1,4 +1,8 @@
-import { TypedMock } from '../utils/test-utils';
+import { getFnMock, TypedMock } from '../utils/test-utils';
 import { EventService } from 'event/event.service';
 
-export default {} as TypedMock<EventService>;
+export default {
+  getEventError: getFnMock<EventService['getEventError']>(),
+  sendErrorEvent: getFnMock<EventService['sendErrorEvent']>(),
+  sendVariableUpdateEvent: getFnMock<EventService['sendVariableUpdateEvent']>(),
+} as TypedMock<EventService>;