+++ test/mocks/variable.service.ts
@@ -1,0 +1,9 @@
+import { getFnMock, TypedMock } from '../utils/test-utils';
+import { VariableService } from 'variable/variable.service';
+
+export default {
+  getAll: getFnMock<VariableService['getAll']>(),
+  toServerVariableSpecification: getFnMock<VariableService['toServerVariableSpecification']>(),
+  findByPath: getFnMock<VariableService['findByPath']>(),
+  getVariableValue: getFnMock<VariableService['getVariableValue']>(),
+} as TypedMock<VariableService>;